+++ src/Processing/Merge/2_merge_ext.jl
@@ -2,9 +2,9 @@ export mseis!, purge!, purge
 
 # Home of all extended merge! methods
 @doc (@doc merge)
-merge(S::SeisData; v::Int64=KW.v) = (U = deepcopy(S); merge!(U, v=v); return U)
-merge!(S::SeisData, U::SeisData; v::Int64=KW.v) = ([append!(getfield(S, f), getfield(U, f)) for f in SeisIO.datafields]; S.n += U.n; merge!(S; v=v))
-merge!(S::SeisData, C::SeisChannel; v::Int64=KW.v) = merge!(S, SeisData(C), v=v)
+merge(S::SeisData; v::Integer=KW.v) = (U = deepcopy(S); merge!(U, v=v); return U)
+merge!(S::SeisData, U::SeisData; v::Integer=KW.v) = ([append!(getfield(S, f), getfield(U, f)) for f in SeisIO.datafields]; S.n += U.n; merge!(S; v=v))
+merge!(S::SeisData, C::SeisChannel; v::Integer=KW.v) = merge!(S, SeisData(C), v=v)
 
 """
     S = merge(A::Array{SeisData,1})
@@ -14,7 +14,7 @@ input data.
 
 See also: merge!
 """
-function merge(A::Array{SeisData,1}; v::Int64=KW.v)
+function merge(A::Array{SeisData,1}; v::Integer=KW.v)
   L::Int64 = length(A)
   n = sum([A[i].n for i = 1:L])
   T = SeisData(n)
@@ -22,10 +22,10 @@ function merge(A::Array{SeisData,1}; v::Int64=KW.v)
   merge!(T, v=v)
   return T
 end
-merge(S::SeisData, U::SeisData; v::Int64=KW.v) = merge(Array{SeisData,1}([S,U]), v=v)
-merge(S::SeisData, C::SeisChannel; v::Int64=KW.v) = merge(S, SeisData(C), v=v)
-merge(C::SeisChannel, S::SeisData; v::Int64=KW.v) = merge(SeisData(C), S, v=v)
-merge(C::SeisChannel, D::SeisChannel; v::Int64=KW.v) = (S = SeisData(C,D); merge!(S, v=v); return S)
+merge(S::SeisData, U::SeisData; v::Integer=KW.v) = merge(Array{SeisData,1}([S,U]), v=v)
+merge(S::SeisData, C::SeisChannel; v::Integer=KW.v) = merge(S, SeisData(C), v=v)
+merge(C::SeisChannel, S::SeisData; v::Integer=KW.v) = merge(SeisData(C), S, v=v)
+merge(C::SeisChannel, D::SeisChannel; v::Integer=KW.v) = (S = SeisData(C,D); merge!(S, v=v); return S)
 
 """
     purge!(S::SeisData)
@@ -36,8 +36,8 @@ Remove empty and duplicated channels in S; alias to merge!(S, purge_only=true)
 
 "Safe" purge to a new SeisData object. Alias to merge(S, purge_only=true)
 """
-purge!(S::T, v::Int64=KW.v) where {T<:GphysData} = merge!(S, purge_only=true, v=v)
-function purge(S::T, v::Int64=KW.v) where {T<:GphysData}
+purge!(S::T, v::Integer=KW.v) where {T<:GphysData} = merge!(S, purge_only=true, v=v)
+function purge(S::T, v::Integer=KW.v) where {T<:GphysData}
   U = deepcopy(S)
   merge!(U, v=v, purge_only=true)
   return U
